derniere version de VAC avec threshold descendant    ajout heuristic last conflict degree ponderee 'q'    ajout singleton vac consistency option 'S'; assez lent    trois bugs identifies

+++ toulbar2/tb2variable.hpp
@@ -79,7 +79,7 @@ public:
     ConstraintList *getConstrs() {return &constrs;}
     int getDegree() {return constrs.getSize();}
     int getTrueDegree();
-    Long getWeightedDegree();
+    double getWeightedDegree();
     DLink<ConstraintLink> *link(Constraint *c, int index);
     void sortConstraints();
     virtual void eliminate() {cout << "variable elimination not implemented!" << endl;};
